backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
bcrypt==4.3.0
Flask-Mail==0.10.0
Flask-Caching==2.1.0
redis==5.0.1
textblob==0.18.0.post0
APScheduler==3.11.0
python-dotenv==1.0.1
pg8000==1.31.2
google-generativeai
